move to dedicated file and enhance readability
